+++ .env-example
@@ -7,5 +7,10 @@ DATABASE_URL=
 GITHUB_ID=
 GITHUB_SECRET=
 
+# Newsletter registration
+LISTMONK_USERNAME=
+LISTMONK_PASSWORD=
+
+# Telegram job offers
 TELEGRAM_TOKEN=
 TELEGRAM_JOBS_CHAT_ID=
